Rooms追加機能。mapで展開。
