--- src/components/Rooms.js
@@ -1,21 +1,0 @@
-import React from "react"
-import styled from "styled-components"
-
-const Rooms = () => {
-    
-    return(
-        <Wrap>
-            <form>
-                <input></input>
-                <button>追加</button>
-            </form>
-            
-        </ Wrap>
-    )
-}
-const Wrap = styled.div`
-    width: 20vw;
-    height: 50vh;
-    background-color: blue;
-`
-export default Rooms
