+++ src/components/rooms/Room.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Room = ({ room }) => {
+  return (
+    <li>
+      {room}
+    </li>
+  )
+}
+
+export default Room
